split frontend json module graph tests
