+++ docs/COMPLETION_AUDIT.md
@@ -283,6 +283,10 @@ and do not assume Phase 4 is ready without evidence.
   diagnostics into
   `tests/integration/cli_build/emit_direct_validation/target_selection.rs`,
   leaving unrelated gated test source skip coverage in the parent module.
+- Frontend JSON integration tests now split AST and symbol module-graph
+  serialization coverage into
+  `tests/integration/cli_build/frontend_json/module_graph.rs`, leaving
+  typed-program and diagnostics JSON checks in the parent module.
 - Generic specialization generated-C assertions now split multi-file import and
   dependency cases into
   `tests/integration/generic_specializations/multifile_generated_c.rs`,
